fix(dev): restrict /_nitro/tasks endpoint to local requests
The dev task runner (/_nitro/tasks and /_nitro/tasks/:name) executes server
tasks with caller-supplied payload and had no access control, while the dev
server binds all interfaces by default (devServer.hostname is unset). Any host
that can reach the dev server could enumerate and run tasks. The VFS viewer
already restricts itself to loopback requests; this applies the same check to
the task runner, sharing the logic in a new isLocalDevRequest helper.
